feat(templates): split AI Chat into conversation and landing page templates
- AI Chat: new conversation view showcasing tool calls, system messages,
  markdown, code blocks, multi-bubble grouping, tokenized text, avatars,
  timestamps, and metadata
- AI Chat (Landing Page): existing landing page template renamed from
  the original AI Chat template (no design changes)
