+++ packages/cli/templates/pages/ai-chat-landing/template.doc.mjs
@@ -1,0 +1,7 @@
+/** @type {import('../../../../core/src/docs-types').TemplateDoc} */
+export const doc = {
+  type: 'page',
+  name: 'AI Chat (Landing Page)',
+  description: 'AI assistant landing page with chat composer, greeting, and category toggles',
+  isReady: true,
+};
